Refactor utility functions and update Installer constructor
Utility functions in the lib/utils module now reside in separate categorised files for better maintenance, which include 'actions.js', 'data.js' and 'fs.js'. Meanwhile, an additional object 'paths' is updated to the Installer class constructor in lib/Installer.js to specify dependency paths and enhance reusability.
