+++ lib/Database.js
@@ -1,5 +1,5 @@
 const mysql = require('mysql2/promise');
-const { getConnectionSettings } = require('./utils');
+const { getConnectionSettings } = require('./data');
 const mysqldump = require('mysqldump');
 const fs = require('fs').promises;
