+++ lib/Installer.js
@@ -14,6 +14,7 @@ class Installer {
    * Initializes a new instance of the constructor.
    *
    * @param {WPMMconfig} config - The configuration object for the constructor.
+   * @param {WPMMpaths} paths - The object containing the paths for the constructor.
    */
   constructor (config, paths) {
     this.config = config;
