+++ lib/Package.js
@@ -2,16 +2,17 @@ const fs = require('fs');
 const path = require('path');
 const { exec } = require('child_process');
 const {
-  getDownloadUrl,
   installNpmPackages,
   installComposer
-} = require('./utils');
-
+} = require('./utils/actions.js');
+const {
+  getDownloadUrl
+} = require('./utils/data.js');
 const {
   downloadFile,
   extractZip,
   renameFolder
-} = require('./utils/fs');
+} = require('./utils/fs.js');
 
 /**
  * Represents a package and provides methods to download and install it.
