+++ lib/Updater.js
@@ -1,4 +1,4 @@
-const { exec } = require('child_process');
+const { exec } = require('node:child_process');
 
 /**
  * Represents the Updater class.
