+++ lib/WpPackage.js
@@ -1,10 +1,10 @@
 const path = require('path');
 const fs = require('fs');
-const { renameFolder } = require('./utils/fs');
+const { renameFolder } = require('./utils/fs.js');
 
-const Package = require('./package');
+const Package = require('./package.js');
 const { replaceDbConstant, replaceDbConstantBool, replaceEmptySalts } = require('./utils/wpConfig.js');
-const { getWordPressDownloadUrl } = require('./utils/index.js');
+const { getWordPressDownloadUrl } = require('./utils/data.js');
 
 /**
  * Represents a WordPress package that can be installed and configured.
